+++ test/data_garrison.sh
@@ -8,7 +8,7 @@ SCRIPT=transload
 PROVIDER="data_garrison"
 STATION_ID="300234065673960"
 USER_ID="300234063581640"
-DATASTORE="/Volumes/ramdisk/datastore/weather"
+DATASTORE="file:///Volumes/ramdisk/datastore/weather"
 DESTINATION="http://192.168.33.77:8080/FROST-Server/v1.0/"
 NOW=$(ruby -e "puts (Time.new).utc.strftime('%FT%T%z')")
 THEN=$(ruby -e "puts (Time.new - (24*3600)).utc.strftime('%FT%T%z')")
@@ -19,14 +19,14 @@ time ruby $SCRIPT get metadata \
     --provider $PROVIDER \
     --station_id $STATION_ID \
     --user_id $USER_ID \
-    --cache $DATASTORE \
+    --database_url $DATASTORE \
     --overwrite
 
 ruby $SCRIPT set metadata \
     --provider $PROVIDER \
     --station_id $STATION_ID \
     --user_id $USER_ID \
-    --cache $DATASTORE \
+    --database_url $DATASTORE \
     --key "latitude" \
     --value "69.158" > /dev/null
 
@@ -34,7 +34,7 @@ ruby $SCRIPT set metadata \
     --provider $PROVIDER \
     --station_id $STATION_ID \
     --user_id $USER_ID \
-    --cache $DATASTORE \
+    --database_url $DATASTORE \
     --key "longitude" \
     --value "-107.0403" > /dev/null
 
@@ -42,7 +42,7 @@ ruby $SCRIPT set metadata \
     --provider $PROVIDER \
     --station_id $STATION_ID \
     --user_id $USER_ID \
-    --cache $DATASTORE \
+    --database_url $DATASTORE \
     --key "timezone_offset" \
     --value "-06:00" > /dev/null
 
@@ -50,19 +50,19 @@ time ruby $SCRIPT put metadata \
     --provider $PROVIDER \
     --station_id $STATION_ID \
     --user_id $USER_ID \
-    --cache $DATASTORE \
+    --database_url $DATASTORE \
     --destination $DESTINATION
 
 time ruby $SCRIPT get observations \
     --provider $PROVIDER \
     --station_id $STATION_ID \
     --user_id $USER_ID \
-    --cache $DATASTORE
+    --database_url $DATASTORE
 
 time ruby $SCRIPT put observations \
     --provider $PROVIDER \
     --station_id $STATION_ID \
     --user_id $USER_ID \
-    --cache $DATASTORE \
+    --database_url $DATASTORE \
     --date "$INTERVAL" \
     --destination $DESTINATION