+++ test/klrs_h_energy.sh
@@ -7,7 +7,7 @@ set -e
 SCRIPT=transload
 PROVIDER="klrs_h_energy"
 STATION_ID="KLRS_Office_Energy"
-DATASTORE="/Volumes/ramdisk/datastore/weather"
+DATASTORE="file:///Volumes/ramdisk/datastore/weather"
 DESTINATION="http://192.168.33.77:8080/FROST-Server/v1.0/"
 
 # KLRS Historical Energy Usage Data Test Run
@@ -20,25 +20,25 @@ time ruby $SCRIPT get metadata \
     --data_path "tmp/klrs-energy/gen_july.xls" \
     --data_path "tmp/klrs-energy/gen_aug.xls" \
     --data_path "tmp/klrs-energy/gen_sep.xls" \
-    --cache $DATASTORE \
+    --database_url $DATASTORE \
     --overwrite
 
 time ruby $SCRIPT put metadata \
     --provider $PROVIDER \
     --station_id $STATION_ID \
-    --cache $DATASTORE \
+    --database_url $DATASTORE \
     --destination $DESTINATION
 
 time ruby $SCRIPT get observations \
     --provider $PROVIDER \
     --station_id $STATION_ID \
-    --cache $DATASTORE
+    --database_url $DATASTORE
 
 # A limited interval is used to prevent this from taking too long during
 # a test.
 time ruby $SCRIPT put observations \
     --provider $PROVIDER \
     --station_id $STATION_ID \
-    --cache $DATASTORE \
+    --database_url $DATASTORE \
     --date "2014-04-28T00:00:00Z/2014-04-29T00:00:00Z" \
     --destination $DESTINATION