+++ test/klrs_h_weather.sh
@@ -7,7 +7,7 @@ set -e
 SCRIPT=transload
 PROVIDER="klrs_h_weather"
 STATION_ID="KLRS_5264"
-DATASTORE="/Volumes/ramdisk/datastore/weather"
+DATASTORE="file:///Volumes/ramdisk/datastore/weather"
 DESTINATION="http://192.168.33.77:8080/FROST-Server/v1.0/"
 
 # KLRS Historical Weather Data Test Run
@@ -25,19 +25,19 @@ time ruby $SCRIPT get metadata \
     --data_path "tmp/klrs-weather/2014/May/TOA5_5264.Health.dat" \
     --data_path "tmp/klrs-weather/2014/September/TOA5_5264.FiveMin.dat" \
     --data_path "tmp/klrs-weather/2014/September/TOA5_5264.Health.dat" \
-    --cache $DATASTORE \
+    --database_url $DATASTORE \
     --overwrite
 
 time ruby $SCRIPT put metadata \
     --provider $PROVIDER \
     --station_id $STATION_ID \
-    --cache $DATASTORE \
+    --database_url $DATASTORE \
     --destination $DESTINATION
 
 time ruby $SCRIPT get observations \
     --provider $PROVIDER \
     --station_id $STATION_ID \
-    --cache $DATASTORE
+    --database_url $DATASTORE
 
 # Upload partial 2012 observations only to save time (approx 4600).
 # (For a local SensorThings API instance, 250,000 observations takes 
@@ -45,6 +45,6 @@ time ruby $SCRIPT get observations \
 time ruby $SCRIPT put observations \
     --provider $PROVIDER \
     --station_id $STATION_ID \
-    --cache $DATASTORE \
+    --database_url $DATASTORE \
     --date "2012-08-01T00:00:00Z/2012-08-07T00:00:00Z" \
     --destination $DESTINATION